Move user page to new file
