+++ webui/pages/index.tsx
@@ -43,14 +43,9 @@ export const Home = (): JSX.Element => {
   const topItems = range(1, 12, 1)
   const bottomItems = range(13, 27, 1)
 
-  const handleSearch = (event: React.MouseEvent<HTMLButtonElement, MouseEvent>) => {
-
-  }
-
-  const handleGoProfile = (event: React.MouseEvent<HTMLButtonElement, MouseEvent>) => {
-    window.location.href = appendBaseURL('user', true);
-    event.preventDefault();
-  }
+  const handleSearch = (
+    event: React.MouseEvent<HTMLButtonElement, MouseEvent>
+  ) => {}
 
   return (
     <>
@@ -66,8 +61,16 @@ export const Home = (): JSX.Element => {
         })}
         <div className={classes.logo}>
           <img src="images/logo.png" />
-          <Button variant="contained" color="primary" onClick={handleSearch}>Search</Button>
-          <Button variant="contained" color="primary" onClick={handleGoProfile}>Go Profile</Button>
+          <Button variant="contained" color="primary" onClick={handleSearch}>
+            Search
+          </Button>
+          <Button
+            variant="contained"
+            color="primary"
+            href={appendBaseURL('user', true)}
+          >
+            Go Profile
+          </Button>
         </div>
         {bottomItems.map((el, index) => {
           return (
